Removed wrapped exports from /src/server/database/
