+++ src/server/database/config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  databaseName : 'layouts',
+  ip: '192.168.90.167',
+  port: '28015'
+};
